backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pyyaml
sqlmodel
python-dotenv
google-genai
